+++ VpNet.sln
@@ -10,6 +10,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		GitVersion.yml = GitVersion.yml
 	EndProjectSection
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VpNet.IntegrationTests", "VpNet.IntegrationTests\VpNet.IntegrationTests.csproj", "{03590718-0BBE-4EEA-B2E6-5C2BA549EF0A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -20,6 +22,10 @@ Global
 		{DBE501A7-0176-473F-B2E8-F4ADA721917F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DBE501A7-0176-473F-B2E8-F4ADA721917F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DBE501A7-0176-473F-B2E8-F4ADA721917F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{03590718-0BBE-4EEA-B2E6-5C2BA549EF0A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{03590718-0BBE-4EEA-B2E6-5C2BA549EF0A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{03590718-0BBE-4EEA-B2E6-5C2BA549EF0A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{03590718-0BBE-4EEA-B2E6-5C2BA549EF0A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
