+++ VpNet/ManagedApi/System/CellCacheManagement/BaseInstanceT_CellCache.cs
@@ -1,10 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using VpNet.Abstract;
 using VpNet.Extensions;
 using VpNet.Interfaces;
 
-namespace VpNet.Abstract
+namespace VpNet.ManagedApi
 {
     /// <summary>
     /// Abtract fully teamplated instance class, providing .NET encapsulation strict templated types to the native C wrapper.
