+++ VpNet/ManagedApi/System/Interfaces/IInstance.cs
